pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apiauth"
version = "0.1.0"
description = "CLI tool for API key and JWT lifecycle management with encrypted local store"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{name = "Revenue Holdings"}]
urls = {Homepage = "https://github.com/Coding-Dev-Tools/apiauth", Repository = "https://github.com/Coding-Dev-Tools/apiauth"}
keywords = ["api-keys", "jwt", "auth", "cli", "security", "key-management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click>=8.1.0",
    "cryptography>=41.0.0",
    "pyjwt>=2.8.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.27.0",
    "freezegun>=1.2.0",
]

[project.scripts]
apiauth = "apiauth.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
